fixed styles on logging details
