+++ src/app/(app)/dashboard/logs/actions.ts
@@ -1,4 +1,3 @@
-
 'use server';
 
 import * as db from '@/lib/db';
@@ -6,6 +5,10 @@ import type { LogEntry } from '@/lib/types';
 
 export async function getRequestLogsAction(): Promise<LogEntry[]> {
   return db.getRequestLogs();
+}
+
+export async function getLogByIdAction(logId: string): Promise<LogEntry | null> {
+  return db.getRequestLogById(logId);
 }
 
 export async function deleteLogEntryAction(logId: string): Promise<void> {
