+++ src/app/(app)/dashboard/logs/page.tsx
@@ -1,554 +1,5 @@
-"use client";
-
-import { useEffect, useState } from "react";
-import { PageShell } from "@/components/layout/PageShell";
-import type { LogEntry, LoggedIntegrationAttempt } from "@/lib/types";
-import { getPlatformFormat, getPlatformFormatDescription } from "@/lib/platform-helpers";
-import {
-  Table,
-  TableBody,
-  TableCell,
-  TableHead,
-  TableHeader,
-  TableRow,
-} from "@/components/ui/table";
-import Link from "next/link";
-import { Badge } from "@/components/ui/badge";
-import { Button } from "@/components/ui/button";
-import { Eye, AlertCircle, CheckCircle, XCircle, HelpCircle, ServerCrash, Trash2, AlertTriangle, RefreshCw, Play, Pause } from "lucide-react";
-import { format, parseISO } from "date-fns";
-import {
-  Dialog,
-  DialogContent,
-  DialogDescription,
-  DialogHeader,
-  DialogTitle,
-  DialogTrigger,
-  DialogClose,
-  DialogFooter,
-} from "@/components/ui/dialog";
-import {
-  AlertDialog,
-  AlertDialogAction,
-  AlertDialogCancel,
-  AlertDialogContent,
-  AlertDialogDescription,
-  AlertDialogFooter,
-  AlertDialogHeader,
-  AlertDialogTitle,
-  AlertDialogTrigger,
-} from "@/components/ui/alert-dialog";
-import { ScrollArea } from "@/components/ui/scroll-area";
-import { Card, CardContent, CardDescription, CardHeader, CardTitle } from "@/components/ui/card";
-import { Prism as SyntaxHighlighter } from 'react-syntax-highlighter';
-import { atomDark } from 'react-syntax-highlighter/dist/esm/styles/prism';
-import { Skeleton } from "@/components/ui/skeleton";
-import { getRequestLogsAction, deleteLogEntryAction, deleteAllLogEntriesAction } from "./actions"; // Updated import
-import { useToast } from "@/hooks/use-toast";
-import { usePolling } from "@/hooks/use-polling";
-
-const StatusIcon = ({ status }: { status: LogEntry['processingSummary']['overallStatus'] }) => {
-  switch (status) {
-    case 'success':
-      return <CheckCircle className="h-5 w-5 text-green-500" />;
-    case 'partial_failure':
-      return <AlertCircle className="h-5 w-5 text-yellow-500" />;
-    case 'total_failure':
-      return <XCircle className="h-5 w-5 text-destructive" />;
-    case 'no_integrations_triggered':
-      return <HelpCircle className="h-5 w-5 text-muted-foreground" />;
-    default:
-      return <HelpCircle className="h-5 w-5 text-muted-foreground" />;
-  }
-};
-
-const IntegrationStatusIcon = ({ status }: { status: LoggedIntegrationAttempt['status'] }) => {
-   switch (status) {
-    case 'success':
-      return <CheckCircle className="h-4 w-4 text-green-500" />;
-    case 'failed_transformation':
-    case 'failed_relay':
-      return <XCircle className="h-4 w-4 text-destructive" />;
-    case 'skipped_disabled':
-    case 'skipped_no_association': // This status might be less relevant now
-      return <HelpCircle className="h-4 w-4 text-muted-foreground" />;
-    default:
-      return <HelpCircle className="h-4 w-4 text-muted-foreground" />;
-  }
-};
+import LogsPageClient from './LogsPageClient';
 
 export default function LogsPage() {
-  const [logs, setLogs] = useState<LogEntry[]>([]);
-  const [isLoading, setIsLoading] = useState(true);
-  const [isDeleting, setIsDeleting] = useState<string | null>(null);
-  const [isDeletingAll, setIsDeletingAll] = useState(false);
-  const [selectedLog, setSelectedLog] = useState<LogEntry | null>(null);
-  const [autoRefreshEnabled, setAutoRefreshEnabled] = useState(true);
-  const [isRefreshing, setIsRefreshing] = useState(false);
-  const { toast } = useToast();
-
-  const fetchLogs = async (showRefreshingIndicator = false) => {
-    if (showRefreshingIndicator) {
-      setIsRefreshing(true);
-    } else {
-      setIsLoading(true);
-    }
-    
-    try {
-      const fetchedLogs = await getRequestLogsAction();
-      // Sorting is handled by DB query (ORDER BY timestamp DESC)
-      setLogs(fetchedLogs);
-    } catch (error) {
-      console.error("Failed to load logs:", error);
-      // Only show toast error if it's not an auto-refresh (to avoid spam)
-      if (!showRefreshingIndicator) {
-        toast({
-          variant: "destructive",
-          title: "Error",
-          description: "Failed to load logs. Please try again.",
-        });
-      }
-    } finally {
-      setIsLoading(false);
-      setIsRefreshing(false);
-    }
-  };
-
-  // Initial load
-  useEffect(() => {
-    fetchLogs();
-  }, []);
-
-  // Auto-refresh polling
-  usePolling(
-    () => fetchLogs(true),
-    {
-      interval: 5000, // 5 seconds
-      enabled: autoRefreshEnabled,
-      immediate: false
-    }
-  );
-
-  const handleManualRefresh = () => {
-    fetchLogs(true);
-  };
-
-  const toggleAutoRefresh = () => {
-    setAutoRefreshEnabled(!autoRefreshEnabled);
-    toast({
-      title: autoRefreshEnabled ? "Auto-refresh disabled" : "Auto-refresh enabled",
-      description: autoRefreshEnabled 
-        ? "Logs will no longer refresh automatically" 
-        : "Logs will refresh every 5 seconds",
-    });
-  };
-
-  const handleViewDetails = (log: LogEntry) => {
-    setSelectedLog(log);
-  };
-
-  const handleDeleteLog = async (logId: string) => {
-    setIsDeleting(logId);
-    try {
-      await deleteLogEntryAction(logId);
-      setLogs(logs.filter(log => log.id !== logId));
-      toast({
-        title: "Log deleted",
-        description: "The log entry has been successfully deleted.",
-      });
-    } catch (error) {
-      console.error("Failed to delete log:", error);
-      toast({
-        title: "Error",
-        description: "Failed to delete the log entry. Please try again.",
-        variant: "destructive",
-      });
-    } finally {
-      setIsDeleting(null);
-    }
-  };
-
-  const handleDeleteAllLogs = async () => {
-    setIsDeletingAll(true);
-    try {
-      await deleteAllLogEntriesAction();
-      setLogs([]);
-      toast({
-        title: "All logs deleted",
-        description: "All log entries have been successfully deleted.",
-      });
-    } catch (error) {
-      console.error("Failed to delete all logs:", error);
-      toast({
-        title: "Error",
-        description: "Failed to delete all log entries. Please try again.",
-        variant: "destructive",
-      });
-    } finally {
-      setIsDeletingAll(false);
-    }
-  };
-  
-  const getStatusBadgeVariant = (status: LogEntry['processingSummary']['overallStatus']) => {
-    switch (status) {
-      case 'success': return 'default';
-      case 'partial_failure': return 'outline';
-      case 'total_failure': return 'destructive';
-      case 'no_integrations_triggered': return 'secondary';
-      default: return 'outline';
-    }
-  };
-  
-  const getIntegrationStatusBadgeVariant = (status: LoggedIntegrationAttempt['status']) => {
-    switch (status) {
-      case 'success': return 'default';
-      case 'failed_transformation':
-      case 'failed_relay': 
-        return 'destructive';
-      case 'skipped_disabled':
-      case 'skipped_no_association':
-        return 'secondary';
-      default: return 'outline';
-    }
-  };
-
-
-  if (isLoading) {
-    return (
-      <PageShell title="Request Logs" description="Inspect incoming requests and their relay status.">
-        <Card className="shadow-lg">
-          <CardHeader>
-            <Skeleton className="h-7 w-1/4" />
-          </CardHeader>
-          <CardContent>
-            <div className="space-y-2">
-              {[...Array(5)].map((_, i) => (
-                <Skeleton key={i} className="h-12 w-full" />
-              ))}
-            </div>
-          </CardContent>
-        </Card>
-      </PageShell>
-    );
-  }
-
-  return (
-    <PageShell
-      title="Request Logs"
-      description={`Inspect recent incoming requests and their relay status. ${autoRefreshEnabled ? ' - Auto-refreshing every 5 seconds' : ''}`}
-    >
-      <Card className="shadow-lg">
-        <CardHeader className="flex flex-row items-center justify-between space-y-0 pb-4">
-          <div>
-            <CardTitle className="flex items-center gap-2">
-              Request Logs
-              {isRefreshing && (
-                <RefreshCw className="h-4 w-4 animate-spin text-muted-foreground" />
-              )}
-            </CardTitle>
-            <CardDescription>
-              Recent webhook requests and their processing status
-              {autoRefreshEnabled && (
-                <span className="text-green-600 ml-1">(Auto-refreshing every 5s)</span>
-              )}
-            </CardDescription>
-          </div>
-          <div className="flex space-x-2">
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={handleManualRefresh}
-              disabled={isRefreshing}
-            >
-              <RefreshCw className={`mr-2 h-4 w-4 ${isRefreshing ? 'animate-spin' : ''}`} />
-              Refresh
-            </Button>
-            <Button
-              variant="outline"
-              size="sm"
-              onClick={toggleAutoRefresh}
-            >
-              {autoRefreshEnabled ? (
-                <>
-                  <Pause className="mr-2 h-4 w-4" />
-                  Pause Auto-refresh
-                </>
-              ) : (
-                <>
-                  <Play className="mr-2 h-4 w-4" />
-                  Start Auto-refresh
-                </>
-              )}
-            </Button>
-            {logs.length > 0 && (
-              <AlertDialog>
-                <AlertDialogTrigger asChild>
-                  <Button 
-                    variant="destructive" 
-                    size="sm"
-                    disabled={isDeletingAll}
-                  >
-                    <Trash2 className="mr-2 h-4 w-4" />
-                    {isDeletingAll ? "Deleting..." : "Delete All"}
-                  </Button>
-                </AlertDialogTrigger>
-              <AlertDialogContent>
-                <AlertDialogHeader>
-                  <AlertDialogTitle>Delete All Log Entries</AlertDialogTitle>
-                  <AlertDialogDescription>
-                    This action cannot be undone. This will permanently delete all log entries from the database.
-                  </AlertDialogDescription>
-                </AlertDialogHeader>
-                <AlertDialogFooter>
-                  <AlertDialogCancel>Cancel</AlertDialogCancel>
-                  <AlertDialogAction 
-                    onClick={handleDeleteAllLogs}
-                    className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
-                  >
-                    Delete All
-                  </AlertDialogAction>
-                </AlertDialogFooter>
-              </AlertDialogContent>
-            </AlertDialog>
-          )}
-          </div>
-        </CardHeader>
-        <CardContent className="pt-6">
-          {logs.length === 0 ? (
-            <div className="flex flex-col items-center justify-center rounded-lg border-2 border-dashed border-muted-foreground/30 p-12 text-center min-h-[300px]">
-              <ServerCrash className="mx-auto h-12 w-12 text-muted-foreground" />
-              <h3 className="mt-4 text-xl font-semibold text-foreground">No logs found</h3>
-              <div className="mt-2 mb-4 text-sm text-muted-foreground">
-                Once requests are made to your API endpoints, they will appear here.
-              </div>
-            </div>
-          ) : (
-            <ScrollArea className="max-h-[calc(100vh-20rem)]">
-              <Table>
-                <TableHeader>
-                  <TableRow>
-                    <TableHead className="w-[50px]">Status</TableHead>
-                    <TableHead>Timestamp</TableHead>
-                    <TableHead>Endpoint Name</TableHead>
-                    <TableHead>Path</TableHead>
-                    <TableHead>Summary</TableHead>
-                    <TableHead className="text-right">Actions</TableHead>
-                  </TableRow>
-                </TableHeader>
-                <TableBody>
-                  {logs.map((log) => (
-                    <TableRow key={log.id}>
-                      <TableCell><StatusIcon status={log.processingSummary.overallStatus} /></TableCell>
-                      <TableCell className="whitespace-nowrap">
-                        {format(parseISO(log.timestamp), "MMM d, yyyy HH:mm:ss")}
-                      </TableCell>
-                      <TableCell>{log.apiEndpointName || "N/A"}</TableCell>
-                      <TableCell className="font-mono text-xs">{log.apiEndpointPath}</TableCell>
-                      <TableCell>
-                        <Badge variant={getStatusBadgeVariant(log.processingSummary.overallStatus)}
-                               className={log.processingSummary.overallStatus === 'success' ? 'bg-green-500 hover:bg-green-600 text-white' : 
-                                          log.processingSummary.overallStatus === 'partial_failure' ? 'bg-yellow-400 hover:bg-yellow-500 text-black' : ''}
-                        >
-                          {log.processingSummary.overallStatus.replace(/_/g, ' ').toUpperCase()}
-                        </Badge>
-                        <span className="ml-2 text-xs text-muted-foreground">{log.processingSummary.message}</span>
-                      </TableCell>
-                      <TableCell className="text-right">
-                        <div className="flex items-center justify-end gap-2">
-                          <Dialog onOpenChange={(open) => !open && setSelectedLog(null)}>
-                            <DialogTrigger asChild>
-                              <Button variant="outline" size="sm" onClick={() => handleViewDetails(log)}>
-                                <Eye className="mr-2 h-4 w-4" /> Details
-                              </Button>
-                            </DialogTrigger>
-                            {selectedLog && selectedLog.id === log.id && (
-                            <DialogContent className="max-w-4xl">
-                              <DialogHeader>
-                                <DialogTitle className="break-all">Log Details: {selectedLog.apiEndpointName || selectedLog.apiEndpointPath}</DialogTitle>
-                                <DialogDescription>
-                                  {format(parseISO(selectedLog.timestamp), "MMM d, yyyy HH:mm:ss zzz")}
-                                </DialogDescription>
-                              </DialogHeader>
-                              <ScrollArea className="max-h-[80vh] pr-6 overflow-x-hidden">
-                                <div className="space-y-6 py-4">
-                                  <Card>
-                                    <CardHeader><CardTitle className="text-lg">Incoming Request</CardTitle></CardHeader>
-                                    <CardContent className="space-y-3 text-sm">
-                                      <div className="break-all"><strong>IP:</strong> {selectedLog.incomingRequest.ip || "N/A"}</div>
-                                      <div className="break-all"><strong>Method:</strong> {selectedLog.incomingRequest.method}</div>
-                                      <div>
-                                        <strong>Headers:</strong>
-                                        <SyntaxHighlighter language="json" style={atomDark} customStyle={{ maxHeight: '200px', overflowY: 'auto', fontSize: '0.8rem' }} className="rounded-md">
-                                          {JSON.stringify(selectedLog.incomingRequest.headers, null, 2)}
-                                        </SyntaxHighlighter>
-                                      </div>
-                                      <div>
-                                        <strong>Raw Body (XML):</strong>
-                                        <SyntaxHighlighter language="xml" style={atomDark} customStyle={{ maxHeight: '200px', overflowY: 'auto', fontSize: '0.8rem' }} className="rounded-md">
-                                          {selectedLog.incomingRequest.bodyRaw}
-                                        </SyntaxHighlighter>
-                                      </div>
-                                    </CardContent>
-                                  </Card>
-
-                                  <Card>
-                                    <CardHeader><CardTitle className="text-lg">Processing Summary</CardTitle></CardHeader>
-                                    <CardContent className="space-y-1 text-sm">
-                                       <div className="flex flex-wrap items-center gap-2">
-                                         <strong>Overall Status:</strong> 
-                                         <Badge 
-                                             variant={getStatusBadgeVariant(selectedLog.processingSummary.overallStatus)}
-                                             className={`${selectedLog.processingSummary.overallStatus === 'success' ? 'bg-green-500 hover:bg-green-600 text-white' : 
-                                                        selectedLog.processingSummary.overallStatus === 'partial_failure' ? 'bg-yellow-400 hover:bg-yellow-500 text-black' : ''}`}
-                                         >
-                                             {selectedLog.processingSummary.overallStatus.replace(/_/g, ' ').toUpperCase()}
-                                         </Badge>
-                                       </div>
-                                      <div>
-                                        <strong>Message:</strong> 
-                                        <div className="break-words mt-1 border border-muted-foreground/20 bg-muted/30 p-2 rounded-md overflow-auto max-w-full">
-                                          <span className="text-sm">{selectedLog.processingSummary.message}</span>
-                                        </div>
-                                      </div>
-                                    </CardContent>
-                                  </Card>
-
-                                  <Card>
-                                    <CardHeader><CardTitle className="text-lg">Integration Attempts ({selectedLog.integrations.length})</CardTitle></CardHeader>
-                                    <CardContent className="space-y-4">
-                                      {selectedLog.integrations.length === 0 ? <div className="text-sm text-muted-foreground">No integrations were attempted.</div> : null}
-                                      {selectedLog.integrations.map((attempt, idx) => (
-                                        <Card key={idx} className="bg-muted/30">
-                                          <CardHeader className="pb-2 pt-4">
-                                            <div className="w-full flex flex-wrap items-start justify-between">
-                                              <div className="flex items-center gap-2 mb-1">
-                                                <IntegrationStatusIcon status={attempt.status}/>
-                                                <span className="break-all font-medium">{attempt.integrationName} ({attempt.platform})</span>
-                                              </div>
-                                              <Badge variant={getIntegrationStatusBadgeVariant(attempt.status)} className={`text-xs ${attempt.status === 'success' ? 'bg-green-500 hover:bg-green-600 text-white' : ''}`}>
-                                                {attempt.status.replace(/_/g, ' ').toUpperCase()}
-                                              </Badge>
-                                            </div>
-                                          </CardHeader>
-                                          <CardContent className="text-xs space-y-2">
-                                            <div>
-                                              <strong>Format:</strong> 
-                                              <span className="ml-2 px-2 py-1 bg-muted rounded-md">{getPlatformFormatDescription(attempt.platform)}</span>
-                                            </div>
-                                            <div>
-                                              <strong>Webhook URL:</strong> 
-                                              <div className="break-all mt-1 border border-muted-foreground/20 bg-muted/30 p-2 rounded-md overflow-auto max-w-full">
-                                                <code className="text-[0.7rem] text-white">{attempt.webhookUrl}</code>
-                                              </div>
-                                            </div>
-                                            {attempt.errorDetails && (
-                                              <div>
-                                                <strong>Error:</strong> 
-                                                <div className="mt-1 border border-red-300/20 bg-red-900/10 p-2 rounded-md overflow-auto max-w-full">
-                                                  <span className="break-words text-red-400">{attempt.errorDetails}</span>
-                                                </div>
-                                              </div>
-                                            )}
-                                            {attempt.outgoingPayload && (
-                                               <div>
-                                                <strong>Outgoing Payload:</strong>
-                                                <div className="mt-2">
-                                                  <SyntaxHighlighter 
-                                                    language={getPlatformFormat(attempt.platform) === 'json' ? 'json' : 'xml'} 
-                                                    style={atomDark} 
-                                                    customStyle={{ 
-                                                      fontSize: '0.75rem',
-                                                      margin: 0,
-                                                      maxHeight: 'none'
-                                                    }}
-                                                    className="rounded-md"
-                                                    showLineNumbers={true}
-                                                  >
-                                                    {attempt.outgoingPayload}
-                                                  </SyntaxHighlighter>
-                                                </div>
-                                               </div>
-                                            )}
-                                            {attempt.responseStatus !== undefined && (
-                                              <div>
-                                                <strong>Response Status:</strong> 
-                                                <span className="ml-2 px-2 py-1 bg-muted rounded-md">{attempt.responseStatus}</span>
-                                              </div>
-                                            )}
-                                            {attempt.responseBody && (
-                                               <div>
-                                                <strong>Response Body:</strong>
-                                                <div className="mt-2">
-                                                  <SyntaxHighlighter 
-                                                    language="text" 
-                                                    style={atomDark} 
-                                                    customStyle={{ 
-                                                      fontSize: '0.75rem',
-                                                      margin: 0,
-                                                      maxHeight: 'none'
-                                                    }}
-                                                    className="rounded-md"
-                                                    showLineNumbers={true}
-                                                  >
-                                                    {attempt.responseBody}
-                                                  </SyntaxHighlighter>
-                                                </div>
-                                               </div>
-                                            )}
-                                          </CardContent>
-                                        </Card>
-                                      ))}
-                                    </CardContent>
-                                  </Card>
-                                </div>
-                              </ScrollArea>
-                              <DialogFooter className="sm:justify-start">
-                                <DialogClose asChild>
-                                  <Button type="button" variant="secondary">Close</Button>
-                                </DialogClose>
-                              </DialogFooter>
-                            </DialogContent>
-                          )}
-                        </Dialog>
-                        <AlertDialog>
-                          <AlertDialogTrigger asChild>
-                            <Button 
-                              variant="outline" 
-                              size="sm"
-                              disabled={isDeleting === log.id}
-                            >
-                              <Trash2 className="mr-2 h-4 w-4" />
-                              {isDeleting === log.id ? "Deleting..." : "Delete"}
-                            </Button>
-                          </AlertDialogTrigger>
-                          <AlertDialogContent>
-                            <AlertDialogHeader>
-                              <AlertDialogTitle>Delete Log Entry</AlertDialogTitle>
-                              <AlertDialogDescription>
-                                This action cannot be undone. This will permanently delete this log entry from the database.
-                              </AlertDialogDescription>
-                            </AlertDialogHeader>
-                            <AlertDialogFooter>
-                              <AlertDialogCancel>Cancel</AlertDialogCancel>
-                              <AlertDialogAction 
-                                onClick={() => handleDeleteLog(log.id)}
-                                className="bg-destructive text-destructive-foreground hover:bg-destructive/90"
-                              >
-                                Delete
-                              </AlertDialogAction>
-                            </AlertDialogFooter>
-                          </AlertDialogContent>
-                        </AlertDialog>
-                        </div>
-                      </TableCell>
-                    </TableRow>
-                  ))}
-                </TableBody>
-              </Table>
-            </ScrollArea>
-          )}
-        </CardContent>
-      </Card>
-    </PageShell>
-  );
+  return <LogsPageClient />;
 }
